+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
  - Added `DrawParameters::ProvokingVertex` and `DrawError::ProvokingVertexNotSupported`.
+ - Added `DrawError::RasterizerDiscardNotSupported`.
  - `DrawParametersBuilder` is now deprecated.
  - `ProgramCreationInput` now has a `outputs_srgb` member.
  - Fixed a bug with offsets of members in arrays.